Add support for arbitrary date formats for netdata alerts.
This adds support to alarm-notify.sh to report dates in alert
notifications using an arbitrary format string.

The current default behavior (just using the default format fromm
`date`, which is simply the ANSI C asctime() default) is preserved.
In the event that the user specified format string doesn't work, the
script will fall back to using the default output of the date command.

THe config file includes comments listing a handful of widely used
internet date formats as examples.

This also removes the explicit quoting around the shell expansion used
to get the date, as it's unneeded (shell expansion results done with $()
is implicitly treated as a single string), and makes it very hard to
get this working properly.