Fix the new date handling in alarm-notify.sh
The original commit unintentionally capitalized the name of the
`date_format` variable in plugins.d/alarm-notify.sh, which caused the
new date handling support to not work at all (but thankfully did not
break any existing functionality).

This corrects that issue, so that the new date handling support works
correctly.