Download public keys by fingerprint in travis
The uid and fingerprint are more legible than the full key, which should
make future changes easier to review.

The keys are used by Travis only to validate that a git tag was signed
with a recognised key. The results of the build are signed separately.

You can confirm the fingerprints in .travis.yml match the ones in
the by running this on the old packaging/gpg.keys:

  gpg --import-options show-only --import packaging/gpg.keys